Resolve "BUG: Incorrect values in "/v1/stakers/{address}/{asset}" endpoint"

Closes #65

See merge request thorchain/midgard!151

+++ api/rest/v1/handlers/handlers.go
@@ -301,9 +301,9 @@ func (h *Handlers) GetStakersAddressData(ctx echo.Context, address string) error
 
 	response := api.StakersAddressDataResponse{
 		PoolsArray:  &assets,
-		TotalEarned: pointy.Int64(int64(details.TotalEarned)),
+		TotalEarned: pointy.Int64(details.TotalEarned),
 		TotalROI:    pointy.Float64(details.TotalROI),
-		TotalStaked: pointy.Int64(int64(details.TotalStaked)),
+		TotalStaked: pointy.Int64(details.TotalStaked),
 	}
 	return ctx.JSON(http.StatusOK, response)
 }
@@ -334,16 +334,16 @@ func (h *Handlers) GetStakersAddressAndAssetData(ctx echo.Context, address strin
 
 		response[i] = api.StakersAssetData{
 			Asset:           helpers.ConvertAssetForAPI(details.Asset),
-			AssetEarned:     pointy.Int64(int64(details.AssetEarned)),
+			AssetEarned:     pointy.Int64(details.AssetEarned),
 			AssetROI:        pointy.Float64(details.AssetROI),
-			AssetStaked:     pointy.Int64(int64(details.AssetStaked)),
+			AssetStaked:     pointy.Int64(details.AssetStaked),
 			DateFirstStaked: pointy.Int64(int64(details.DateFirstStaked)),
-			PoolEarned:      pointy.Int64(int64(details.PoolEarned)),
+			PoolEarned:      pointy.Int64(details.PoolEarned),
 			PoolROI:         pointy.Float64(details.PoolROI),
-			PoolStaked:      pointy.Int64(int64(details.PoolStaked)),
-			RuneEarned:      pointy.Int64(int64(details.RuneEarned)),
+			PoolStaked:      pointy.Int64(details.PoolStaked),
+			RuneEarned:      pointy.Int64(details.RuneEarned),
 			RuneROI:         pointy.Float64(details.RuneROI),
-			RuneStaked:      pointy.Int64(int64(details.RuneStaked)),
+			RuneStaked:      pointy.Int64(details.RuneStaked),
 			StakeUnits:      pointy.Int64(int64(details.StakeUnits)),
 		}
 	}
